+++ src/api/fetchPosts.js
@@ -5,6 +5,7 @@ export async function fetchPostsByLikes() {
   try {
     const posts = await pb.collection('posts').getFullList({
       sort: '-likedNum', // 좋아요 수 기준 내림차순 정렬
+      autoCancel: false,
     });
     return posts;
   } catch (error) {
